Move Configuration to compas
