+++ docs/examples/02_description_models/files/02_robot_artist_grasshopper_panda.ghx
@@ -497,7 +497,7 @@ robot = st[robot_key]
                 <chunk name="Container">
                   <items count="11">
                     <item name="CodeInput" type_name="gh_string" type_code="10">import math
-from compas_fab.robots import Configuration
+from compas.robots import Configuration
 
 if robot:
     joint_values = [a1, a2, a3, a4, a5, a6, a7]
@@ -2651,7 +2651,7 @@ W = 3</item>
                   <items count="11">
                     <item name="CodeInput" type_name="gh_string" type_code="10">import math
 from compas.robots import Joint
-from compas_fab.robots import Configuration
+from compas.robots import Configuration
 
 if robot:
     joint_values = [j1]
