+++ docs/examples/03_backends_ros/files/05_plan_motion.py
@@ -1,6 +1,6 @@
 import math
 from compas.geometry import Frame
-from compas_fab.robots import Configuration
+from compas.robots import Configuration
 from compas_fab.backends import RosClient
 
 with RosClient() as client:
