+++ docs/examples/05_backends_pybullet/files/02_forward_kinematics.py
@@ -1,6 +1,6 @@
 import compas_fab
+from compas.robots import Configuration
 from compas_fab.backends import PyBulletClient
-from compas_fab.robots import Configuration
 
 with PyBulletClient() as client:
     urdf_filename = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
