+++ src/compas_fab/backends/ros/backend_features/move_it_plan_cartesian_motion.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 
 import itertools
 
+from compas.robots import Configuration
 from compas.utilities import await_callback
 
 from compas_fab.backends.interfaces import PlanCartesianMotion
@@ -19,7 +20,6 @@ from compas_fab.backends.ros.messages import MultiDOFJointState
 from compas_fab.backends.ros.messages import Pose
 from compas_fab.backends.ros.messages import RobotState
 from compas_fab.backends.ros.service_description import ServiceDescription
-from compas_fab.robots import Configuration
 from compas_fab.robots import JointTrajectory
 
 __all__ = [
