+++ src/compas_fab/backends/vrep/helpers.py
@@ -2,11 +2,11 @@ import socket
 
 from compas.geometry import Frame
 from compas.geometry import matrix_from_frame
+from compas.robots import Configuration
 
 from compas_fab.backends.exceptions import BackendError
 from compas_fab.backends.vrep.remote_api import vrep
 
-from compas_fab.robots import Configuration
 
 __all__ = [
     'VrepError',
