+++ src/compas_fab/robots/__init__.py
@@ -17,7 +17,6 @@ Main classes
 
     Robot
     RobotSemantics
-    Configuration
     Tool
     Duration
     Wrench
