+++ src/compas_fab/robots/trajectory.py
@@ -3,10 +3,10 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from compas.robots.configuration import Configuration
+from compas.robots.configuration import FixedLengthList
 from compas_fab.robots import AttachedCollisionMesh
 from compas_fab.robots.time_ import Duration
-from compas_fab.robots.configuration import Configuration
-from compas_fab.robots.configuration import FixedLengthList
 
 __all__ = [
     'JointTrajectoryPoint',
